(LOG-11874) - Corrigido teste unitário
